+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_build.cs
@@ -4,7 +4,8 @@ public static partial class SessionBaseExtensionMethods
 {
 
     /// <summary>
-    /// A <see cref="Pith.SessionBase"/> extension method that trim compress load convert export.
+    /// A <see cref="Pith.SessionBase"/> extension method that trims, loads, optionally converts to
+    /// binary, and compresses the script to the deploy file.
     /// </summary>
     /// <remarks>   2025-04-15. </remarks>
     /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
@@ -16,7 +17,7 @@ public static partial class SessionBaseExtensionMethods
     /// <param name="session">      The session. </param>
     /// <param name="scriptInfo">   Information describing the script. </param>
     /// <param name="folderPath">   Full pathname of the folder file. </param>
-    public static void TrimCompressLoadConvertExport( this Pith.SessionBase session, IScriptInfo scriptInfo, string folderPath )
+    public static void TrimCompressLoadConvertExport( this Pith.SessionBase session, ScriptInfoBase scriptInfo, string folderPath )
     {
         if ( session is null ) throw new ArgumentNullException( nameof( session ) );
         if ( scriptInfo is null ) throw new ArgumentNullException( nameof( scriptInfo ) );
@@ -25,14 +26,18 @@ public static partial class SessionBaseExtensionMethods
 
         string fromFilePath = Path.Combine( folderPath, scriptInfo.BuiltFileName );
         string trimmedFilePath = Path.Combine( folderPath, scriptInfo.TrimmedFileName );
-        string compressedFilePath = Path.Combine( folderPath, $"{scriptInfo.TrimmedFileName}{ScriptInfoBase.SelectScriptFileExtension( ScriptFileFormats.Compressed )}" );
+        string deployFilePath = Path.Combine( folderPath, scriptInfo.DeployFileName );
         if ( System.IO.File.Exists( fromFilePath ) )
         {
             SessionBaseExtensionMethods.TraceLastAction( $"Trimming script file '{fromFilePath}' to '{trimmedFilePath}'" );
             fromFilePath.TrimScript( trimmedFilePath, true );
 
-            SessionBaseExtensionMethods.TraceLastAction( $"Compressing trimmed script file to '{compressedFilePath}'" );
-            trimmedFilePath.CompressScriptFile( compressedFilePath, overWrite: true );
+            if ( scriptInfo.DeployFileFormat.HasFlag( ScriptFileFormats.Compressed )
+                && !scriptInfo.DeployFileFormat.HasFlag( ScriptFileFormats.Binary ) )
+            {
+                SessionBaseExtensionMethods.TraceLastAction( $"Compressing trimmed script file to '{deployFilePath}'" );
+                trimmedFilePath.CompressScriptFile( deployFilePath, overWrite: true );
+            }
         }
         else
             throw new FileNotFoundException( fromFilePath );
@@ -52,20 +57,25 @@ public static partial class SessionBaseExtensionMethods
         // run the script to ensure the code works.
         session.RunScript( scriptInfo.Title, scriptInfo.VersionGetterElement );
 
-        string deployFileName = $"{scriptInfo.DeployFileTitle}{ScriptInfoBase.SelectScriptFileExtension( ScriptFileFormats.Binary )}";
-        string binaryFilePath = Path.Combine( folderPath, deployFileName );
-
-        session.ExportScript( scriptInfo.Title, binaryFilePath, true );
-
-        deployFileName = $"{scriptInfo.DeployFileTitle}{ScriptInfoBase.SelectScriptFileExtension( ScriptFileFormats.Compressed | ScriptFileFormats.Binary )}";
-        compressedFilePath = Path.Combine( folderPath, deployFileName );
-
-        SessionBaseExtensionMethods.TraceLastAction( $"Compressing binary script file to '{compressedFilePath}'" );
-        binaryFilePath.CompressScriptFile( compressedFilePath, overWrite: true );
+        if ( scriptInfo.DeployFileFormat.HasFlag( ScriptFileFormats.Binary ) )
+        {
+            if ( scriptInfo.DeployFileFormat.HasFlag( ScriptFileFormats.Compressed ) )
+            {
+                // export and compress the script.
+                SessionBaseExtensionMethods.TraceLastAction( $"Fetching and compressing binary script to '{deployFilePath}'" );
+                session.CompressScript( scriptInfo.Title, deployFilePath, true );
+            }
+            else
+            {
+                // if naked binary, export tot he deploy file name.
+                SessionBaseExtensionMethods.TraceLastAction( $"Fetching binary script to '{deployFilePath}'" );
+                session.ExportScript( scriptInfo.Title, deployFilePath, true );
+            }
+        }
     }
 
     /// <summary>
-    /// A <see cref="Pith.SessionBase"/> extension method that loads the scripts and saves it to non-
+    /// A <see cref="Pith.SessionBase"/> extension method that loads the script from the deploy file and saves it to non-
     /// volatile memory.
     /// </summary>
     /// <remarks>   2025-04-15. </remarks>
@@ -78,20 +88,21 @@ public static partial class SessionBaseExtensionMethods
     /// <param name="session">          The session. </param>
     /// <param name="scriptInfo">       Information describing the script. </param>
     /// <param name="folderPath">       Full pathname of the folder file. </param>
-    public static void LoadSaveToNvm( this Pith.SessionBase session, IScriptInfo scriptInfo, string folderPath )
+    public static void LoadSaveToNvm( this Pith.SessionBase session, ScriptInfoBase scriptInfo, string folderPath )
     {
         if ( session is null ) throw new ArgumentNullException( nameof( session ) );
         if ( scriptInfo is null ) throw new ArgumentNullException( nameof( scriptInfo ) );
 
         if ( !session.IsDeviceOpen ) throw new InvalidOperationException( $"{nameof( session )} is not open." );
 
-        string fromFilePath = Path.Combine( folderPath, scriptInfo.DeployFileName );
+        // set the deploy file path.
+        string deployFilePath = Path.Combine( folderPath, scriptInfo.DeployFileName );
 
         // delete the script if it exists.
         session.DeleteScript( scriptInfo.Title );
 
-        SessionBaseExtensionMethods.TraceLastAction( $"Importing script from {scriptInfo.DeployFileFormat} '{fromFilePath}' file" );
-        session.ImportScript( scriptInfo.Title, fromFilePath );
+        SessionBaseExtensionMethods.TraceLastAction( $"Importing script from {scriptInfo.DeployFileFormat} '{deployFilePath}' file" );
+        session.ImportScript( scriptInfo.Title, deployFilePath );
 
         session.RunScript( scriptInfo.Title, scriptInfo.VersionGetterElement );